Fixed End Game Implementation
The previous implementation made it very hard for the user to see where exactly they messed up as the end game screen would pop up so quickly. Now by added a setTimeout function the user will be able to see where they made their mistake. Additionally I've changed the gameOver logic to where it only has to be 1 line of code instead of 4 by putting it inside of a function. The -40% hit detection is now 100% perfect.